Added FolderPathLength script
Searches for files with too long path length.
Move the files in the folder which you want to scan the content (or
directly on the C:\ drive), then run start.bat to generate the txt file.
Note: Make sure you have enough permissions to generate the txt file in
the same folder as the script!